Implemented 20x20x20 grid of Rubik's cubes
Created rubik20.cpp to render a 20x20x20 grid of Rubik's cubes.
These cubes are rendered in exactly the same way as the cube from
rubik.cpp.

Bugfixes:
In drawRubik() I called glTranslatef() at the very beginning of
the function. Which this is sufficient when drawing a single cube,
this causes additional cubes to be placed relative to the previous
cube instead of relative to the origin. This was resolved by
calling glPushMatrix() at the beginning of drawRubik() at then
glPopMatrix() at the end.